Include update_engine_sideload in the recovery.
update_engine_sideload is used to sideload an Android OTA .zip file from
the recovery image via sdcard or "adb sideload".

This patch also marks the postinstall step in system as optional since
it is not *required* to mark an update as successful and it also
doesn't run from recovery.

Bug: 27178350
Change-Id: Ibdb82ad4cf9acb1d392319ebe4300b4fdd2cbb44
Signed-off-by: spezi77 <spezi77@gmx.com>